+++ CHANGELOG.md
@@ -9,6 +9,7 @@
 ### Changes
 - Prohibit bioformats and vips from reading mrxs directly ([#1328](../../pull/1328))
 - Handle Python 3.12 deprecating utcnow ([#1331](../../pull/1331))
+- Turn down logging about annotation ACLs ([#1332](../../pull/1332))
 
 ## 1.25.0